refactor(p1): extract WebMountNavigationTools.kt — 6 navigation/state tools + applicableProfileJson
Closes Navigation block for WebMount Group A. Six factories take the standard
WebMountDeps wrapper; wm_open + wm_state additionally take three more deps
(profileRegistry, cookieProvider, manager) so they can build the `applicable_profile`
JSON payload. Other four (extract/get/back/forward) need only deps.

applicableProfileJson moves into NavigationTools.kt as a file-private top-level
function with explicit deps params — only callers are wm_open + wm_state in this
file, so no need for a wider abstraction.

safeUrlPreview moves with wm_open (sole caller). safeEvalPreview stays in
coordinator because wm_eval hasn't migrated yet (Group B).

Coordinator collapses 6 `private val xxxTool = Tool(...)` + 1 helper into
6 lazy delegators. WebMountPrimitiveTools.kt: 1414 → 1071 lines (-24%).

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
